Refined some comments, and move orc_cast_float_test to orc_cast_test
